agent: Unify agent logic to crate `agent-utils`
