+++ agent/utils/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "agent-utils"
+version = "0.0.1"
+edition = "2018"
+publish = false
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+agent-common = { version = "0.0.1", path = "../../agent/agent-common" }
+aws-config = "0.47"
+aws-types = "0.47"
+aws-sdk-sts = "0.17"
+aws-smithy-types = "0.47"
+base64 = "0.13.0"
+env_logger = "0.9"
+log = "0.4"
+model = { version = "0.0.1", path = "../../model" }
+resource-agent = { version = "0.0.1", path = "../../agent/resource-agent" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+snafu = "0.7"
+test-agent = { version = "0.0.1", path = "../../agent/test-agent" }
