+++ agent/utils/src/constants.rs
@@ -1,0 +1,4 @@
+use log::LevelFilter;
+
+pub const DEFAULT_REGION: &str = "us-west-2";
+pub const DEFAULT_AGENT_LEVEL_FILTER: LevelFilter = LevelFilter::Info;
