+++ bottlerocket/agents/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 agent-common = { version = "0.0.1", path = "../../agent/agent-common" }
+agent-utils = { version = "0.0.1", path = "../../agent/utils" }
 bottlerocket-types = { version = "0.0.1", path = "../types" }
 async-trait = "0.1"
 aws-config = "0.47"
