+++ bottlerocket/agents/src/bin/eks-resource-agent/main.rs
@@ -7,7 +7,7 @@ Provides
 mod eks_provider;
 
 use crate::eks_provider::{EksCreator, EksDestroyer};
-use bottlerocket_agents::init_agent_logger;
+use agent_utils::init_agent_logger;
 use resource_agent::clients::{DefaultAgentClient, DefaultInfoClient};
 use resource_agent::error::AgentResult;
 use resource_agent::{Agent, BootstrapData, Types};
