+++ bottlerocket/agents/src/constants.rs
@@ -1,0 +1,3 @@
+pub const DEFAULT_TASK_DEFINITION: &str = "testsys-bottlerocket-aws-default-ecs-smoke-test-v1";
+pub const TEST_CLUSTER_KUBECONFIG_PATH: &str = "/local/test-cluster.kubeconfig";
+pub const E2E_REPO_CONFIG_PATH: &str = "/local/e2e-repo-config.yaml";
